+++ packages/client/modules/pages/Page.tsx
@@ -4,6 +4,7 @@ import type {Page_page$key} from '../../__generated__/Page_page.graphql'
 import type {Page_viewer$key} from '../../__generated__/Page_viewer.graphql'
 import {useIsAuthenticated} from '../../components/IsAuthenticatedProvider'
 import {usePageProvider} from '../../hooks/usePageProvider'
+import {cn} from '../../ui/cn'
 import {DatabaseEditor} from './DatabaseEditor'
 import {PageEditor} from './PageEditor'
 import {PageHeader} from './PageHeader'
@@ -56,7 +57,12 @@ export const Page = (props: Props) => {
   return (
     <div className='relative flex flex-col items-center bg-white'>
       {isPublic ? <PageHeaderPublic /> : <PageHeader pageRef={page} />}
-      <div className='relative flex min-h-screen w-full max-w-[960px] justify-center bg-white pt-28 pb-10 print:pt-0 print:caret-transparent'>
+      <div
+        className={cn(
+          'relative flex min-h-screen w-full justify-center bg-white pt-28 pb-10 print:pt-0 print:caret-transparent',
+          isDatabase ? 'max-w-9/10' : 'max-w-[960px]'
+        )}
+      >
         {synced &&
           (isDatabase ? (
             <DatabaseEditor viewerRef={viewer} isEditable={isEditable} provider={provider} />